MyBatisPlus从3.5.3更新到3.5.3.1，修复调用lambdaQuery().one()方法查询为null是的报错问题。完成后台分类模块接口开发

+++ backend/src/main/java/com/my/blog/pojo/po/UploadFileRef.java
@@ -19,7 +19,7 @@ public class UploadFileRef {
     // 业务类型
     private Integer bizType;
     // 业务id
-    private Integer bizId;
+    private Long bizId;
     // 文件id
     private Long uploadFileId;
     // 状态
